chore: 백엔드 폴더구조 변경 및 환경 설정
- 백엔드 폴더 구조를 변경
- .env를 비롯한 백엔드 환경설정
