+++ server/package.json
@@ -21,6 +21,7 @@
     "prettier": "2.1.2"
   },
   "dependencies": {
+    "dotenv": "^8.2.0",
     "koa": "^2.13.0",
     "koa-bodyparser": "^4.3.0",
     "koa-logger": "^3.2.1",
