+++ server/src/config/server.js
@@ -1,0 +1,11 @@
+const dotenv = require('dotenv');
+
+const envFound = dotenv.config();
+
+if (envFound.error) {
+  throw new Error(" Couldn't find .env file ");
+}
+
+module.exports = {
+  port: process.env.PORT,
+};
