+++ server/src/loaders/index.js
@@ -1,0 +1,8 @@
+const koa = require('./koa');
+
+const server = async (app) => {
+  await koa(app);
+  console.log('🍟 Express Connect! 🍟');
+};
+
+module.exports = server;
